pyproject.toml
[tool.poetry]
name = "drive"
version = "0.1.0"
description = "An open-source alternative for already existing propietary personal file storage solutions."
authors = ["kydrid"]
license = "MIT"
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.11"
fastapi = "^0.95.2"
python-box = "^7.0.1"
tomli = "^2.0.1"
uvicorn = "^0.22.0"


[tool.poetry.group.dev.dependencies]
black = "^23.3.0"
isort = "^5.12.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
